backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
click==8.0.1
fastapi==0.88.0
h11==0.12.0
pip==22.3.1
pydantic==1.10.2
setuptools==58.0.4
starlette==0.22.0
typing-extensions== 4.4.0
uvicorn==0.15.0
passlib==1.7.4
pytest==7.2.0
requests==2.28.1
email-validator==1.1.3
python-jose==3.3.0
python-multipart==0.0.5
tenacity==8.0.1
httpx==0.23.1
pymongo==4.3.3
motor==3.1.1
pytest-asyncio==0.20.3
beanie==1.17.0